+++ NetcodeSamples/Assets/Samples/HelloNetcode/3_Advanced/XXb_ServerSideAnimation/GhostAnimationGraph_InAir.cs
@@ -13,13 +13,12 @@ using Unity.NetCode.Hybrid;
 
 namespace Samples.HelloNetcode.Hybrid
 {
+#if !UNITY_DISABLE_MANAGED_COMPONENTS
     public struct InAirAnimationData : IComponentData
     {
         [GhostField(Quantization=1000)] public float Phase;
         [GhostField(Quantization=1000)] public float aimPitch;
     }
-
-#if !UNITY_DISABLE_MANAGED_COMPONENTS
     public class InAirGhostPlayableBehaviour : GhostPlayableBehaviour
     {
         GhostAnimationController m_controller;